Workaround for multiple definitions errors for export_TFnc() with Visual Studio
